server/requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
click==8.1.8
fastapi==0.115.11
gunicorn==23.0.0
h11==0.14.0
idna==3.10
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
uuid==1.30
uvicorn==0.34.0
websockets==15.0
